+++ Source/Framework/Core/Components/TeCRenderable.h
@@ -126,7 +126,7 @@ namespace te
         float GetUseForZPrepass() const { return _internal->GetUseForZPrepass(); }
 
         /** @copydoc Renderable::SetLayer */
-        void SetLayer(UINT64 layer) { _internal->SetLayer(layer); }
+        void SetLayer(UINT32 layer) { _internal->SetLayer(layer); }
 
         /** @copydoc Renderable::GetLayer */
         UINT64 GetLayer() const { return _internal->GetLayer(); }
